+++ services/madoc-ts/src/types/schemas/create-canvas.ts
@@ -1,10 +1,4 @@
-// import { CanvasNormalized } from '@hyperion-framework/types';
-
 export type CreateCanvas = {
   canvas: any;
-  // canvas: Partial<CanvasNormalized>;
-
   thumbnail?: string;
-
-  local_source?: string;
 };