+++ services/madoc-ts/src/types/schemas/create-manifest.ts
@@ -1,8 +1,6 @@
 // import { ManifestNormalized } from '@hyperion-framework/types';
 
 export type CreateManifest = {
-  // manifest: Partial<ManifestNormalized>;
   manifest: any;
-  local_source?: string;
   taskId?: string;
 };