+++ services/madoc-ts/src/utility/get-viewing-direction.ts
@@ -1,0 +1,24 @@
+import { ViewingDirection } from '@hyperion-framework/types';
+
+export function getViewingDirection(viewingDirection: string | number): ViewingDirection {
+  let viewDir: ViewingDirection = 'left-to-right';
+  switch (viewingDirection) {
+    case 'left-to-right':
+    case 0:
+      viewDir = 'left-to-right';
+      break;
+    case 'right-to-left':
+    case 1:
+      viewDir = 'right-to-left';
+      break;
+    case 'top-to-bottom':
+    case 2:
+      viewDir = 'top-to-bottom';
+      break;
+    case 'bottom-to-top':
+    case 3:
+      viewDir = 'bottom-to-top';
+      break;
+  }
+  return viewDir;
+}